requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.3
pyxirr>=0.8.0
marshmallow==3.20.1